backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.1
httpx==0.27.0

langgraph==0.2.45

# OpenAI
openai==1.54.4

# Database (with srv support for MongoDB Atlas)
pymongo[srv]==4.10.1

# Data processing
pandas==2.2.3

# PDF extraction
pypdf==4.3.1
pdfplumber==0.11.4
Pillow==11.0.0

# CLIP model for embeddings
torch==2.5.1
transformers==4.46.3
